dp-trigger-controller/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob==12.19.1
requests==2.31.0
azure-identity>=1.15.0
